Added a primitive modal to add scripts from editor.
Please note that, at the state of codebase on this commit, At the end of the program a crash occurs if you add a script to a component and don't delete/remove it beforehand. This is because right now ModuleScriptingSystem does not handle the deletion of scripts created by editor. This feature will be added in the following commits.